% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.data.frame.bsimPred}
\alias{as.data.frame.bsimPred}
\title{Extract the Result Data.frame}
\usage{
\method{as.data.frame}{bsimPred}(x, ...)
}
\arguments{
\item{x}{An object of class "bsimPred" which is the result of \code{predict.bsim}.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
data.frame of prediction information.
}
\description{
Extracts fitted result that contains posterior mean/median prediction of response variable, and
standard error, credible interval of each data point in data.frame.
}
\examples{
\donttest{
data("DATA1")
simdata2 <- data.frame(DATA1$X, y = DATA1$y)
fit_one  <- BayesSIM(y ~ ., data = simdata2,
                    niter = 5000, nburnin = 1000, nchain = 1)

pred <- predict(fit_one, se.fit = TRUE, interval = "credible",level = 0.8)
results <- as.data.frame(pred)
summary(results)
}

}
\seealso{
\code{\link{predict.bsim}}
}
