% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSIM-package.R, R/zzz.R
\docType{package}
\name{BayesSIM-package}
\alias{BayesSIM-package}
\title{BayesSIM: Integrated Interface of Bayesian Single Index Models using 'nimble'}
\description{
Provides tools for fitting Bayesian single index models with flexible choices of priors for both the index and the link function. The package implements model estimation and posterior inference using efficient MCMC algorithms built on the 'nimble' framework, allowing users to specify, extend, and simulate models in a unified and reproducible manner. The following methods are implemented in the package: Antoniadis et al. (2004) \url{https://www.jstor.org/stable/24307224}, Wang (2009) \doi{10.1016/j.csda.2008.12.010}, Choi et al. (2011) \doi{10.1080/10485251003768019}, Dhara et al. (2019) \doi{10.1214/19-BA1170}, McGee et al. (2023) \doi{10.1111/biom.13569}.

Provides tools for fitting Bayesian single index models with flexible choices of priors for both the index and the link function. The package implements model estimation and posterior inference using efficient MCMC algorithms built on the 'nimble' framework, allowing users to specify, extend, and simulate models in a unified and reproducible manner. The following methods are implemented in the package: Antoniadis et al. (2004) \url{https://www.jstor.org/stable/24307224}, Wang (2009) \doi{10.1016/j.csda.2008.12.010}, Choi et al. (2011) \doi{10.1080/10485251003768019}, Dhara et al. (2019) \doi{10.1214/19-BA1170}, McGee et al. (2023) \doi{10.1111/biom.13569}.
}
\author{
\strong{Maintainer}: Seowoo Jung \email{jsw1347@ewha.ac.kr}

Authors:
\itemize{
  \item Eun-kyung Lee \email{lee.eunk@ewha.ac.kr}
}

}
\keyword{internal}
