% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_gp.R
\name{plot.bsimGp}
\alias{plot.bsimGp}
\title{Traceplot for bsimGp Results}
\usage{
\method{plot}{bsimGp}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bsimGp}.}

\item{...}{Further arguments passed to \code{\link{plot}}.}
}
\value{
Traceplots for MCMC samples are displayed.
By default, only the index vector and error variance are included in the summary.
If a spike-and-slab prior model, indexstar is parameterized to original index vector.
}
\description{
Provides traceplot for objects of class \code{bsimGp}, corresponding to a
single-index model with a Gaussian process link function.
Z
}
