% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_spline.R
\name{plot.bsimSpline}
\alias{plot.bsimSpline}
\title{Traceplot for bsimSpline Results}
\usage{
\method{plot}{bsimSpline}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bsimSpline}.}

\item{...}{Further arguments passed to \code{\link{plot}}.}
}
\value{
Traceplots for MCMC samples are displayed.
By default, only the index vector and error variance are included in the summary.
}
\description{
Provides traceplot for objects of class \code{bsimSpline}, corresponding to a
single-index model with a Gaussian process link function.
}
