% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.bsimPred}
\alias{summary.bsimPred}
\alias{print.summary.bsimPred}
\title{Summarize Predictions}
\usage{
\method{summary}{bsimPred}(object, ...)

\method{print}{summary.bsimPred}(x, digits = 3, ...)
}
\arguments{
\item{object}{An object of class "bsimPred" which is the result of \code{predict}.}

\item{...}{Further arguments passed.}

\item{x}{A summary output of class "bsimPred".}

\item{digits}{The minimum number of significant digits to be printed.}
}
\value{
Metrics are including RMSE, MAE, Mean bias, and coverage. In terms of coverage, it is only available
when the "bsimPred" includes interval, corresponding to \code{interval = "credible"} options for \code{predict} function.
It indicates mean coverage rate of data points whether confidence intervals include true response values.
In this case, all chains are combined for the computation.
}
\description{
Provides a \code{summary} for \code{BayesSIM} prediction.
}
\details{
Performance of prediction with point predicted values.
}
\examples{
\donttest{
simdata2 <- data.frame(DATA1$X, y = DATA1$y)

fit_one <- BayesSIM(y ~ ., data = simdata2,
                    niter = 5000, nburnin = 1000, nchain = 1)

# Prediction with 95\% credible interval at new data
newx <- data.frame(X1 = rnorm(10), X2 = rnorm(10), X3 = rnorm(10), X4 = rnorm(10))
pred <- predict(fit_one, newdata = newx, interval = "credible", level = 0.95)
summary(pred)
plot(pred)

}
}
