% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.bsimSetup}
\alias{summary.bsimSetup}
\alias{print.summary.bsimSetup}
\title{Summarize \code{BayesSIM_setup}}
\usage{
\method{summary}{bsimSetup}(object, ...)

\method{print}{summary.bsimSetup}(x, digits = 1, ...)
}
\arguments{
\item{object}{A fitted object of \code{BayesSIM_setup}.}

\item{...}{Further arguments passed.}

\item{x}{A summary output of \code{BayesSIM_setup}.}

\item{digits}{The minimum number of significant digits to be printed.}
}
\value{
An object of class \code{summary.bsimSetup}, which is a list containing:
\item{modelName}{The name of the fitted model.}
\item{n}{The number of observations.}
\item{p}{The number of predictors.}
\item{prior}{A list of prior settings used in the model.}
\item{samplingOptions}{A list of MCMC sampling options, including \code{niter}, \code{nburnin}, \code{thin}, and \code{nchain}.}
\item{model}{The compiled nimble model object.}
}
\description{
Provides a \code{summary} for \code{BayesSIM_setup}.
}
\details{
Summarize the model setup information of an object created by
a setup function (e.g., \code{\link{BayesSIM_setup}}),
including the model name, data dimensions, prior settings, and sampling options.
}
\examples{
\donttest{
simdata2 <- data.frame(DATA1$X, y = DATA1$y)

models <- BayesSIM_setup(y ~ ., data = simdata2)
summary(models)

Ccompile <- compileModelAndMCMC(models)
nimSampler <- getSampler(Ccompile)
initList <- getInit(models)
mcmc.out <- runMCMC(nimSampler, niter = 5000, nburnin = 1000, thin = 1,
                    nchains = 1, setSeed = TRUE, inits = initList,
                    summary = TRUE, samplesAsCodaMCMC = TRUE)

# "fit_split" becomes the class of "bsim".
fit_split <- as_bsim(models, mcmc.out)

}

}
