% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.data.frame.bsimPred}
\alias{as.data.frame.bsimPred}
\title{Extract the Result Data.frame}
\usage{
\method{as.data.frame}{bsimPred}(x, ...)
}
\arguments{
\item{x}{An object of class "bsimPred" which is the result of \code{predict.bsim}.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
data.frame of prediction information.
}
\description{
Extracts fitted result that contains posterior mean/median prediction of response variable, and
standard error, credible interval of each data point in data.frame.
}
\seealso{
\code{\link{predict.bsim}}
}
