% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getModelDef}
\alias{getModelDef}
\title{Get Definition of the Model}
\usage{
getModelDef(object)
}
\arguments{
\item{object}{A fitted object of \code{BayesSIM} or individual model.}
}
\value{
BUGS code of the model definition.
}
\description{
Functions for identifying definition of the \pkg{nimble} model.
}
\details{
The function that contain Bayes SIM model structure in \pkg{nimble}. This function is for advanced users.
}
\examples{
simdata2 <- data.frame(DATA1$X, y = DATA1$y)
models <- bsFisher_setup(y ~ ., data = simdata2)
# Get model definition
getModelDef(models)

}
\seealso{
\code{\link{getVarMonitor}}
}
