% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_bspline_ver3.R
\docType{data}
\name{SplineState}
\alias{SplineState}
\title{Additional Functions for Spline}
\format{
An object of class \code{list} of length 1.
}
\usage{
SplineState
}
\description{
Additional Functions for Spline
}
\keyword{datasets}
